refactor: refactor how query params for filtering are used
refactor: org imports
