+++ src/api/query-hooks/useJobsHistoryQuery.ts
@@ -1,6 +1,12 @@
+import { durationOptions } from "@flanksource-ui/components/JobsHistory/Filters/JobHistoryDurationDropdown";
+import { jobHistoryDefaultDateFilter } from "@flanksource-ui/components/JobsHistory/Filters/JobsHistoryFilters";
+import { JobHistory } from "@flanksource-ui/components/JobsHistory/JobsHistoryTable";
+import { resourceTypeMap } from "@flanksource-ui/components/SchemaResourcePage/SchemaResourceEditJobsTab";
+import useTimeRangeParams from "@flanksource-ui/ui/TimeRangePicker/useTimeRangeParams";
 import { useQuery, UseQueryOptions } from "@tanstack/react-query";
-import { JobHistory } from "../../components/JobsHistory/JobsHistoryTable";
-import { getJobsHistory } from "../services/jobsHistory";
+import { useMemo } from "react";
+import { useSearchParams } from "react-router-dom";
+import { getJobsHistory, GetJobsHistoryParams } from "../services/jobsHistory";
 
 type Response =
   | { error: Error; data: null; totalEntries: undefined }
@@ -17,70 +23,57 @@ export function useJobsHistoryQuery(
 ) {
   return useQuery<Response, Error>(
     ["jobs_history", pageIndex, pageSize],
-    () => getJobsHistory(pageIndex, pageSize),
+    () => getJobsHistory({ pageIndex, pageSize }),
     options
   );
 }
 
-type QueryInputType = {
-  pageIndex: number;
-  pageSize: number;
-  resourceType?: string;
-  resourceId?: string;
-  name?: string;
-  status?: string;
-  sortBy?: string;
-  sortOrder?: string;
-  startsAt?: string;
-  endsAt?: string;
-  duration?: number;
-};
-
 export function useJobsHistoryForSettingQuery(
-  {
+  options?: UseQueryOptions<Response, Error>,
+  resourceId?: string,
+  tableName?: keyof typeof resourceTypeMap
+) {
+  const { timeRangeAbsoluteValue } = useTimeRangeParams(
+    jobHistoryDefaultDateFilter
+  );
+
+  const [searchParams] = useSearchParams();
+  const pageIndex = parseInt(searchParams.get("pageIndex") ?? "0");
+  const pageSize = parseInt(searchParams.get("pageSize") ?? "150");
+  const name = searchParams.get("name") ?? "";
+  const sortBy = searchParams.get("sortBy") ?? "";
+  const sortOrder = searchParams.get("sortOrder") ?? "desc";
+  const status = searchParams.get("status") ?? "";
+  const duration = searchParams.get("runDuration") ?? undefined;
+  const durationMillis = duration
+    ? durationOptions[duration].valueInMillis
+    : undefined;
+  const startsAt = timeRangeAbsoluteValue?.from ?? undefined;
+  const endsAt = timeRangeAbsoluteValue?.to ?? undefined;
+  const resourceType = useMemo(() => {
+    if (!tableName) {
+      return undefined;
+    }
+    return resourceTypeMap[tableName];
+  }, [tableName]);
+
+  const params = {
     pageIndex,
     pageSize,
     resourceType,
-    resourceId,
     name,
     status,
     sortBy,
     sortOrder,
     startsAt,
     endsAt,
-    duration
-  }: QueryInputType,
-  options?: UseQueryOptions<Response, Error>
-) {
+    duration: durationMillis,
+    resourceId
+  } satisfies GetJobsHistoryParams;
+
   return useQuery<Response, Error>(
-    [
-      "jobs_history",
-      pageIndex,
-      pageSize,
-      resourceType,
-      resourceId,
-      name,
-      status,
-      sortBy,
-      sortOrder,
-      startsAt,
-      endsAt,
-      duration
-    ],
-    () =>
-      getJobsHistory(
-        pageIndex,
-        pageSize,
-        resourceType,
-        resourceId,
-        name,
-        status,
-        sortBy,
-        sortOrder,
-        startsAt,
-        endsAt,
-        duration
-      ),
+    ["jobs_history", params],
+    () => getJobsHistory(params),
     options
   );
 }
